refactor: DictionaryProvideService 를 최적화
