+++ src/main/java/toy/subscribe/dictionary/service/DictionaryProvideService.java
@@ -1,8 +1,26 @@
 package toy.subscribe.dictionary.service;
 
+import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import toy.subscribe.configs.http.log.repository.HttpLogRepository;
 import toy.subscribe.configs.http.wrapper.HttpResponseWrapper;
+import toy.subscribe.dictionary.dto.DictionaryResponseDto;
+import toy.subscribe.dictionary.repository.DictionaryRepository;
 
-public interface DictionaryProvideService {
-    HttpResponseWrapper provideDictionaryWrapper(Pageable pageable, String enWord, String krWord);
+@Service
+@RequiredArgsConstructor
+public class DictionaryProvideService {
+    private final DictionaryRepository feedBoardRepository;
+    private final HttpLogRepository requestLogRepository;
+
+    @Transactional(readOnly = true)
+    public HttpResponseWrapper<DictionaryResponseDto> provideDictionaryWrapper(Pageable pageable, String enWord, String krWord) {
+        return HttpResponseWrapper.<DictionaryResponseDto>builder()
+                                  .pages(feedBoardRepository.findPage(pageable, enWord, krWord))
+                                  .visitorsOfDay(requestLogRepository.findCumulativeVisitors())
+                                  .visitorsOfReduce(requestLogRepository.findDau())
+                                  .build();
+    }
 }
